Remove FormikStepper
